test(compose): relocate caption-kwarg test under _composition mirror (audit PS-204)
