+++ tests/figrecipe/_composition/test__compose.py
@@ -1275,3 +1275,28 @@ class TestComposeRawImages:
             is_image_file as _is_image_file,
         )
         assert not _is_image_file(Path("test.py"))
+
+
+def test_compose_caption_kwarg_persists_on_record(tmp_path):
+    """Card compose-caption-kwarg: compose(caption=...) attaches caption
+    text to the composed figure's FigureRecord."""
+    # Arrange
+    fig1, ax1 = fr.subplots()
+    ax1.plot([1, 2, 3], [1, 4, 9], id="line1")
+    recipe1 = tmp_path / "fig1.yaml"
+    fr.save(fig1, recipe1, validate=False, verbose=False)
+
+    fig2, ax2 = fr.subplots()
+    ax2.bar([1, 2, 3], [3, 1, 4], id="bar1")
+    recipe2 = tmp_path / "fig2.yaml"
+    fr.save(fig2, recipe2, validate=False, verbose=False)
+
+    # Act
+    composed, _ = fr.compose(
+        layout=(1, 2),
+        sources={(0, 0): recipe1, (0, 1): recipe2},
+        caption="Combined plot of line and bar.",
+    )
+
+    # Assert
+    assert composed.record.caption == "Combined plot of line and bar."
